% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_classes.R
\docType{class}
\name{CSmosquitoSp.obj}
\alias{CSmosquitoSp.obj}
\alias{CSmosquitoSp.obj-class}
\title{Class "CSmosquitoSp.obj"}
\description{
Result from \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} stored as a \code{\linkS4class{CSmosquitoSp.obj}} object.
Slots for \code{CSmosquitoSp.obj} objects include: (1) the reported points from the mosquito DataBase;
(2) the AOI spatial polygons containing results from the spatial analysis of the mosquito DataBase.
}
\section{Slots}{

\describe{
\item{\code{DB}}{Character. Type of the imported Citizen Science mosquito DataBase}

\item{\code{points}}{SpatialPointsDataFrame. Report spatial points used for the analysis}

\item{\code{aoi}}{SpatialPolygonsDataFrame. Spatial polygons of the selected Area Of Interest containing the results from the spatial analysis}

\item{\code{aoi_name}}{Character. Name of the selected Area Of Interest}

\item{\code{start_date}}{Date. Start day selected for the analysis}

\item{\code{end_date}}{Date. End day selected for the analysis}

\item{\code{method}}{Character. Method used for the analysis (currently 'RC' and 'CM' are supported)}
}}

\section{Objects from the class}{
 Objects are created by calls to \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}}.
}

\seealso{
\code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} \code{\link[CSmosquitoSp]{CSmosquitoSp.plot}} \code{\link[CSmosquitoSp]{CSmosquitoSp.export}}
}
\author{
Federico Filipponi
}
