% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_analysis.R
\name{CSmosquitoSp.analysis}
\alias{CSmosquitoSp.analysis}
\title{CSmosquitoSp spatial analysis}
\usage{
CSmosquitoSp.analysis(input_db, aoi, start_date, end_date, method = "CM",
  full_clean = FALSE, aoi_name = NULL, ...)
}
\arguments{
\item{input_db}{DataFrame. Input CSmosquitoSp DataBase imported using function \code{\link[CSmosquitoSp]{CSmosquitoSp.import}}.}

\item{aoi}{Character. Area Of Interest where to perform the spatial analysis.}

\item{start_date}{Character. Input start date, either an object of class \code{\linkS4class{Date}} or a character object with format 'YYYY-MM-DD'.}

\item{end_date}{Character. Input end date, either an object of class \code{\linkS4class{Date}} or a character object with format 'YYYY-MM-DD'.}

\item{method}{Character. Method to be used for the analysis.
The following options are supported: 'CM' (Class Median), 'RC' (ReClass) (default to 'CM').}

\item{full_clean}{Logical. Wheter the full cleaning of the input DataBase
should be performed prior the analysis.}

\item{aoi_name}{Character. Name for the Area Of Interest when 'aoi' argument is supplied as external file.}

\item{...}{Additional arguments to be passed through to function \code{\link{CSmosquitoSp.analysis}}}
}
\value{
The function returns an object of class \code{\linkS4class{CSmosquitoSp.obj}} containing the following slots:
\itemize{
  \item points Report spatial points used for the analysis
  \item  aoi Spatial polygons of the selected Area Of Interest containing the results from the spatial analysis
  \item  aoi_name Name of the selected Area Of Interest
  \item  start_date Start day selected for the analysis
  \item  end_date End day selected for the analysis
  \item  method Method used for the analysis (currently 'CM' and 'RC' are supported)
}
}
\description{
This function perform spatial analysis for a selected AOI
using data from the CSmosquitoSp DataBase
}
\examples{
\dontrun{
# perform CSmosquitoSp spatial analysis
for the metropolitan area of the city of Rome
CSMS_roma <- CSmosquitoSp.analysis(input_db=input_db,
aoi="Roma", start_date="2017-09-01", end_date="2017-09-01")

# perform CSmosquitoSp spatial analysis using the 'full_clean' option
for the metropolitan area of the city of Rome
CSMS_roma_full_clean <- CSmosquitoSp.analysis(input_db=input_db,
aoi="Roma", start_date="2017-09-01", end_date="2017-09-01", full_clean=TRUE)
}

}
\seealso{
\code{\link[CSmosquitoSp]{CSmosquitoSp.plot}} \code{\link[CSmosquitoSp]{CSmosquitoSp.export}} \code{\link[CSmosquitoSp]{CSmosquitoSp.import}}
}
\author{
Federico Filipponi, Mattia Manica
}
\keyword{CSmosquitoSp}
\keyword{analysis}
\keyword{spatial}
