% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_export.R
\name{CSmosquitoSp.export}
\alias{CSmosquitoSp.export}
\title{CSmosquitoSp export}
\usage{
CSmosquitoSp.export(x, output_path, points = FALSE, kml = FALSE,
  sensitive = FALSE)
}
\arguments{
\item{x}{Input object of class \code{\linkS4class{CSmosquitoSp.obj}}}

\item{output_path}{Character. Path where maps are saved on disk.}

\item{points}{Logical. When set to TRUE export points used for the analysis also in 'shapefile' format (default to FALSE).}

\item{kml}{Logical. When set to TRUE results are also exported in 'kml' format (default to FALSE).}

\item{sensitive}{Logical. When set to TRUE export the sensitive information in records (default to FALSE).}
}
\value{
The function export the results generated using function \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} to disk.
Output vector file format are 'shapefile' and 'kml' (optional).
}
\description{
This function export results computed using \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} function
a specific Area Of Interest and a selected temporal range.
}
\examples{
\dontrun{
# perform CSmosquitoSp spatial analysis
for the metropolitan area of the city of Rome
CSMS_roma <- CSmosquitoSp.analysis(input_db=input_db, aux_path=aux_path,
aoi="Roma", start_date="2017-09-01", end_date="2017-09-01")

# export result
CSmosquitoSp.export(x=CSMS_result, output_path="C:/output")

# export result also in kml format
CSmosquitoSp.export(x=CSMS_result, output_path="C:/output", kml=TRUE)

# export also points used for the analysis
CSmosquitoSp.export(x=CSMS_result, output_path="C:/output", points=TRUE)
}

}
\seealso{
\code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} \code{\link[CSmosquitoSp]{CSmosquitoSp.plot}}
}
\author{
Federico Filipponi
}
\keyword{CSmosquitoSp}
\keyword{export}
