% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_import.R
\name{CSmosquitoSp.import}
\alias{CSmosquitoSp.import}
\title{CSmosquitoSp import}
\usage{
CSmosquitoSp.import(x, type = NULL)
}
\arguments{
\item{x}{Character. Input path where CSmosquitoSp DataBase in CSV format is stored on disk.}

\item{type}{Character. Type of the imported Citizen Science mosquito DataBase (at present only 'ZanzaMapp' type is supported)}
}
\value{
The function return a DataFrame containing the CSmosquitoSp DataBase
}
\description{
This function import CSmosquitoSp DataBase to be used for spatial analysis using function \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}}
}
\examples{
\dontrun{
# set input path
input_path <- normalizePath(path=system.file("external/CSmosquitoSp_database.csv"),
winslash="/", mustWork=TRUE)
# import mosquito DataBase of type 'ZanzaMapp'
input_db <- CSmosquitoSp.import(x=input_path, type="ZanzaMapp")

###
# directly input mosquito DataBase file path in the function
input_db <- CSmosquitoSp.import(x="C:/CSmosquitoSp_database.csv", type="ZanzaMapp")
}
}
\seealso{
\code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} \code{\link[CSmosquitoSp]{CSmosquitoSp.plot}} \code{\link[CSmosquitoSp]{CSmosquitoSp.export}}
}
\author{
Federico Filipponi
}
\keyword{CSmosquitoSp}
\keyword{import}
