% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_plot.R
\name{CSmosquitoSp.plot}
\alias{CSmosquitoSp.plot}
\title{CSmosquitoSp plot}
\usage{
CSmosquitoSp.plot(x, field = "VAL_PMFA",
  maptitle = "Perceived mosquito abundance", lang = "ENG",
  filename = NULL, caption_logo = FALSE, ...)
}
\arguments{
\item{x}{Input object of class \code{\linkS4class{CSmosquitoSp.obj}}}

\item{field}{Character. Fields to be mapped from the 'aoi' slot of object of class \code{\linkS4class{CSmosquitoSp.obj}}}

\item{maptitle}{Character. 'Map title.}

\item{lang}{Character. Language to use for map footer ('ENG' or 'ITA').}

\item{filename}{Character. Plot result to speified file path.}

\item{caption_logo}{Logical. Add the caption logo (currently available only for 'ZanzaMapp' DataBase) to the map (Default to FALSE).}

\item{...}{Additional arguments to be passed through to function \code{\link{CSmosquitoSp.plot}}}
}
\value{
The function return a plot for the selected field computed from \code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} within the Area Of Interest
}
\description{
This function create spatial maps for a selected AOI using the results
generated from spatial analysis of CSmosquitoSp data
}
\examples{
\dontrun{
# perform CSmosquitoSp spatial analysis
for the metropolitan area of the city of Rome
CSMS_roma <- CSmosquitoSp.analysis(input_db=input_db, aux_path=aux_path,
aoi="Roma", start_date="2017-09-01", end_date="2017-09-30")

# plot result
CSmosquitoSp.plot(x=CSMS_roma, field="VAL_PMFA")
}

}
\seealso{
\code{\link[CSmosquitoSp]{CSmosquitoSp.analysis}} \code{\link[CSmosquitoSp]{CSmosquitoSp.export}}
}
\author{
Federico Filipponi
}
\keyword{CSmosquitoSp}
\keyword{plot}
