% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,CSmosquitoSp.obj-method}
\title{CSmosquitoSp methods}
\usage{
\S4method{show}{CSmosquitoSp.obj}(object)
}
\arguments{
\item{object}{CSmosquitoSp.obj. Object of class \code{\linkS4class{CSmosquitoSp.obj}}}
}
\description{
This function define methods for the CSmosquitoSp package for class \code{\linkS4class{CSmosquitoSp.obj}}
a specific Area Of Interest and a selected temporal range.
}
\author{
Federico Filipponi
}
