% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSmosquitoSp_valw.R
\name{val_w}
\alias{val_w}
\title{CSmosquitoSp val_w}
\usage{
val_w(x, y)
}
\arguments{
\item{x}{Numeric. Value x}

\item{y}{Numeric. Value y}
}
\description{
Function for weighted mean
}
\author{
Federico Filipponi
}
